wip: renames and typos
