+++ src/apps/renderer/localize/locales/en.json
@@ -329,7 +329,7 @@
       "no-remote-connection": "We could not connect to Internxt servers, we'll try again later",
       "cannot-get-current-listings": "We could not get the status of your current files, we'll try again later",
       "base-directory-does-not-exist": "Unable to find your sync folder",
-      "insuficient-permision-accessing-base-directory": "Internxt App does not have permission to access your sync folder",
+      "insufficient-permission-accessing-base-directory": "Internxt App does not have permission to access your sync folder",
       "cannot-access-base-directory": "We could not access your local folder",
       "cannot-access-tmp-directory": "We could not access your  local folder",
       "unknown": "An unknown error ocurred while trying to sync your files",
