+++ src/apps/renderer/localize/locales/es.json
@@ -328,7 +328,7 @@
       "no-remote-connection": "No hemos podido conectar con los servidores de Internxt, lo intentaremos más tarde",
       "cannot-get-current-listings": "No pudimos obtener el estado de tus archivos actuales, lo intentaremos más tarde",
       "base-directory-does-not-exist": "No se pudo encontrar su carpeta de sincronización",
-      "insuficient-permision-accessing-base-directory": "Internxt App no tiene permiso para acceder a su carpeta de sincronización",
+      "insufficient-permission-accessing-base-directory": "Internxt App no tiene permiso para acceder a su carpeta de sincronización",
       "cannot-access-base-directory": "No hemos podido acceder a su carpeta local",
       "cannot-access-tmp-directory": "No hemos podido acceder a su carpeta local",
       "unknown": "Error desconocido al intentar sincronizar sus archivos",
