+++ src/apps/renderer/messages/backups/backups-fatal-errors.ts
@@ -1,4 +1,4 @@
-import { ProcessFatalErrorName } from '../../../shared/types';
+import { ProcessFatalErrorName } from '../../../../shared/issues/VirtualDriveIssue';
 import processFatalErrors from '../process-fatal-error';
 
 const messages: Partial<Record<ProcessFatalErrorName, string>> = {
