+++ src/apps/renderer/messages/process-error.ts
@@ -12,7 +12,7 @@ export const shortMessages: ProcessErrorMessages = {
   UNKNOWN: 'issues.short-error-messages.unknown',
   FILE_TOO_BIG: 'issues.short-error-messages.file-too-big',
   FILE_NON_EXTENSION: 'issues.short-error-messages.file-non-extension',
-  DUPLICATED_NODE: 'issues.short-error-messages.duplicated-node',
+  DUPLICATED: 'issues.short-error-messages.duplicated-node',
 };
 
 export const longMessages: ProcessErrorMessages = {
@@ -26,5 +26,5 @@ export const longMessages: ProcessErrorMessages = {
   UNKNOWN: 'issues.error-messages.unknown',
   FILE_TOO_BIG: 'issues.error-messages.file-too-big',
   FILE_NON_EXTENSION: 'issues.error-messages.file-non-extension',
-  DUPLICATED_NODE: 'issues.error-messages.duplicated-node',
+  DUPLICATED: 'issues.error-messages.duplicated-node',
 };
