+++ src/apps/shared/types.ts
@@ -1,127 +1,4 @@
-import { Readable } from 'stream';
 import { AppErrorName } from '../../shared/issues/AppIssue';
-
-export interface FileSystem {
-  /**
-   * The kind of filesystem, it's emitted
-   * in some fs events
-   */
-  kind: FileSystemKind;
-
-  /**
-   * Returns the listing of the current files
-   * in this FileSystem
-   */
-  getCurrentListing(): Promise<{
-    listing: Listing;
-    readingMetaErrors: ReadingMetaErrorEntry[];
-  }>;
-
-  /**
-   * Renames a file in the FileSystem
-   * @param oldName
-   * @param newName
-   */
-  renameFile(oldName: string, newName: string): Promise<void>;
-
-  /**
-   * Deletes a file in the FileSystem,
-   * doesn't throw if the file doesn't exist anymore
-   * @param name
-   */
-  deleteFile(name: string): Promise<void>;
-
-  /**
-   * Pulls a file from other FileSystem into this FileSystem,
-   * overwriting it if already exists
-   * @param name
-   * @param source
-   * @param progressCallback
-   */
-  pullFile(
-    name: string,
-    source: Source,
-    progressCallback: FileSystemProgressCallback
-  ): Promise<number | void>;
-
-  /**
-   * Checks if a folder exists in the filesystem
-   * @param name
-   */
-  existsFolder(name: string): Promise<boolean>;
-
-  /**
-   * Deletes a folder in the filesystem
-   * doesn't throw if the folder doesn't exist anymore
-   * @param name
-   */
-  deleteFolder(name: string): Promise<void>;
-
-  /**
-   * Returns an object source that contains
-   * anything that another filesystem would need
-   * to pull it
-   * @param name
-   * @param progressCallback
-   */
-  getSource(
-    name: string,
-    progressCallback: FileSystemProgressCallback
-  ): Promise<Source>;
-
-  /**
-   * Check critical resources of this filesystem
-   * and throw an error if it's not operative
-   */
-  smokeTest(): Promise<void>;
-}
-
-export type FileSystemKind = 'LOCAL' | 'REMOTE';
-
-export type ReadingMetaErrorEntry = {
-  name: string;
-  errorName: ProcessErrorName;
-  errorDetails: ErrorDetails;
-};
-
-export type FileSystemProgressCallback = (progress: number) => void;
-
-export type Source = {
-  stream: Readable;
-  modTime: number;
-  size: number;
-};
-
-/**
- * Represents a list of files, each with
- * its modTime that is set as seconds since epoch
- * and size in bytes
- *
- * The name of each file can be namespaced by
- * his ancestors such as: folderA/folderB/fileName
- * It cannot start or end with "/"
- */
-export type Listing = Record<string, { modtime: number; size: number }>;
-
-export type ProcessFatalErrorName =
-  | 'NO_INTERNET'
-  | 'NO_REMOTE_CONNECTION'
-  | 'CANNOT_ACCESS_BASE_DIRECTORY'
-  | 'BASE_DIRECTORY_DOES_NOT_EXIST'
-  | 'INSUFICIENT_PERMISION_ACCESSING_BASE_DIRECTORY'
-  | 'CANNOT_ACCESS_TMP_DIRECTORY'
-  | 'CANNOT_GET_CURRENT_LISTINGS'
-  | 'UNKNOWN';
-
-export class ProcessFatalError extends Error {
-  details: ErrorDetails;
-
-  constructor(name: ProcessFatalErrorName, details: ErrorDetails) {
-    super();
-    this.name = name;
-    this.details = details;
-  }
-}
 
 export type ProcessErrorName =
   // File or folder does not exist
@@ -152,7 +29,7 @@ export type ProcessErrorName =
   | 'UNKNOWN'
 
   // Duplicated node path
-  | 'DUPLICATED_NODE';
+  | 'DUPLICATED';
 
 export class ProcessError extends Error {
   details: ErrorDetails;
