+++ src/shared/issues/VirtualDriveIssue.ts
@@ -11,3 +11,13 @@ export type VirtualDriveIssue = {
   node: string;
   errorName: ProcessErrorName;
 };
+
+export type ProcessFatalErrorName =
+  | 'NO_INTERNET'
+  | 'NO_REMOTE_CONNECTION'
+  | 'CANNOT_ACCESS_BASE_DIRECTORY'
+  | 'BASE_DIRECTORY_DOES_NOT_EXIST'
+  | 'INSUFFICIENT_PERMISSION_ACCESSING_BASE_DIRECTORY'
+  | 'CANNOT_ACCESS_TMP_DIRECTORY'
+  | 'CANNOT_GET_CURRENT_LISTINGS'
+  | 'UNKNOWN';
